requirements.txt
# Core ML dependencies
transformers>=4.30.0
datasets>=2.12.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
torch>=2.2.0  # Updated for compatibility
torchvision>=0.17.0  # Updated for compatibility
torchaudio>=2.2.0  # Updated for compatibility
numpy>=1.24.0
pandas>=2.0.0
mistralai>=0.0.12  # Mistral AI client - Pinned to a working version

# RAG and Language Processing
langchain>=0.0.267
python-dotenv>=1.0.0

# Web Application
flask>=2.0.0
flask-cors>=3.0.10
streamlit==1.22.0
requests>=2.28.0

# Utilities
tqdm>=4.65.0
matplotlib>=3.7.0
plotly>=5.13.0
scikit-learn>=1.0.0

# Development
pytest>=7.0.0
black>=22.0.0
isort>=5.12.0
rouge_score>=0.1.2
